Thao1

+++ AzCare-APIServices/src/main/java/com/project/AzCar/Service/Deposit/ICarbankService.java
@@ -11,9 +11,11 @@ public interface ICarbankService {
 	List<Cardbank> getListCardBank();
 
 	List<Cardbank> getListCardBankAdmin();
+
 	Cardbank saveCardbank(Cardbank c);
 
 	Cardbank findCardbankbyId(int id);
 
-	Cardbank findCardbankByUserId(int id);
+	Cardbank findCardbankByUserId(long id);
+
 }
